CXO-109 - added NSEW to coordinates, corrected coordinates validation, added elevation for coordinates

+++ copl/com.elster.jupiter.util/src/main/java/com/elster/jupiter/util/geo/WorldCoordinate.java
@@ -11,6 +11,6 @@ public interface WorldCoordinate extends Serializable {
 
     Integer getMinutes();
 
-    BigDecimal getSeconds();
+    Integer getSeconds();
 
 }
